Deciders. MD5 and Timestamp.
Defaults to MD5. It will only rebuild if the MD5 signature of one of the sources
changes in between rebuilds.

This garbage is completely untested. Also the web server will refresh the page
and then serve the old version which is a little redundant.

This code won't work: isBuilderCurrent doesn't understand when one of the
sources is another Builder. It is also a destrictive operation, so if it's
called twice it will determine that none of the sources have changed on the
second time.