Default Makefile support; new Makefile format
This builds a default Makefile that contains all of the rules of the previous
build system but specified in such a way that builders can be swapped out for
other ones in a user Makefile. This paves the way to start constructing
pipelines, e.g. JSX -> Modulr -> Uglify.

The current DefaultMakefile is incomplete and has some information commented for
as it becomes available.